Add nix flake + direnv config, improve readme
The introduction of a nix flake makes it easier to build and flash the firmware
with the Nix package manager.

Besides being tweaked to describe Nix usage instructions, more information has
been added to the readme to make building the monocle more accessible to people
who do not use VSCode.