add an RTT logging module
This is permitting to send RTT logs from Python for debugging purposes.

It does not make sense to have it when the user does not have access to
an RTT debugger, and therefore it is not compiled in by default, but it
can be convenient to have it nearby so that it can quickly be added to
the Makefile for tracing some bug.