sprites: merged with display.py
It works only the first time for now.
The data can only be sent once, but the
description and placement can be sent multiple times.
Maybe it needs to be flushed completely?
It seems to be possible to update the raw data, but not
to append anything.
This might still work with what we might want to do,
but the user-facing API will be impacted.