Reorder the execution of _splashscreen.py and _mountfs.py
If a file on the filesystem is called `display.py`, and this file triggers a
bug, such as a hard fault, then `_splashscreen.py` would be importing this
dangerous file from the filesystem, and the Monocle would be stuck in a boot
loop.

Executing `_splashscreen.py` before the filesystem is mounted avoids breaking
the Monocle in case of a name collision with an user-provided file,
protecting the boot sequence from bugs elsewhere in the firmware.